nix: 2.16->2.15; home/fish/atuin: remove up key binding
nix-plugins built for a newer nix versions than pkgs.nix makes using
pkgs.nix impossible, and overriding pkgs.nix globally breaks some
derivations, so I have to do per-app overrides (couldn't find a way to
revert it for a single app for some reason). This reverts a nix-plugins
update to 2.16 so I can build it for 2.15 and escape the "symbol not
found" hell.

also override nixVersions.unstable to be 2.15 as well

finally, use custom code for atuin integration because I hate the way it
normally hijacks my up key